Removed Ruby version
Its performance was already bad, but a benchmark on an incomplete port
of the TLG6 reader proved that it was going to decrease even more.

It simply made no sense to keep the Ruby implementation around if it's
execution time was 30x slower (TLG5) and 250x slower (TLG6)... even
after micro-optimizations.